requirements.txt
fastapi
uvicorn
langchain
openai
faiss-cpu
line-bot-sdk
python-dotenv
